pyproject.toml
[project]
name = "docbot"
version = "0.1.0"
description = "WhatsApp-based appointment management system for single-doctor practice"
readme = "README.md"
authors = [
    { name = "KanaparthySaiSreekar", email = "kanapasai@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-jose[cryptography]>=3.3.0",
    "authlib>=1.3.0",
    "httpx>=0.27.0",
    "itsdangerous>=2.2.0",
    "aiosqlite>=0.20.0",
    "python-multipart>=0.0.9",
    "jinja2>=3.1.0",
    "python-json-logger>=2.0.7",
    "tzdata>=2025.3",
    "google-api-python-client>=2.189.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.2.4",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
]

[build-system]
requires = ["uv_build>=0.9.7,<0.10.0"]
build-backend = "uv_build"
